refactor: shared 레이어의 도메인 파일들 features 레이어로 이동
